Modele - BugFix 1

+++ app/models/Chart.scala
@@ -1,3 +1,3 @@
 package models
 
-case class Chart(id: Int, user: Int, product: Int, count: Int)
+case class Chart(id: Int, user: Int, product: Long, count: Int)
